单例模式自动释放，配置类增加功能，架构微调，代码规范化

+++ keyword_suggestion/server/src/online/task_queue.cc
@@ -12,7 +12,9 @@ TaskQueue::TaskQueue(size_t queue_size)
 
 bool TaskQueue::IsEmpty() const { return !queue_.size(); }
 
-bool TaskQueue::IsFull() const { return queue_.size() == queue_size_; }
+bool TaskQueue::IsFull() const {
+  return static_cast<int>(queue_.size()) == queue_size_;
+}
 
 void TaskQueue::Push(const ElemType &element) {
   MutexLockGuard auto_lock(mutex_);
